[gem] Use OVShape in open_vino_execute_calculator (#4484)
Summary: OpenVinoExecuteCalculator should just set the request input
shape from the input cpu packet.

Relevant Issues: n/a

Type of change: /kind bugfix

Test Plan: Tested on skaffold gem

GitOrigin-RevId: 0f1f3b2d01a5faad446c315dd38468713e57dc7a

+++ src/gem/calculators/plugin/cpu_tensor/tokens_to_tensor_calculator.cc
@@ -49,7 +49,7 @@ Status TokensToTensorCalculator::ProcessImpl(mediapipe::CalculatorContext* cc,
   const auto& tokens = cc->Inputs().Tag(kTokensTag).Get<std::vector<int>>();
 
   GML_ASSIGN_OR_RETURN(auto tensor, exec_ctx->TensorPool()->GetTensor(
-                                        {1, static_cast<int>(tokens.size())}, DataType::INT64));
+                                        {1, static_cast<int>(tokens.size())}, DataType::INT32));
 
   std::memcpy(tensor->data(), tokens.data(), sizeof(int) * tokens.size());
